python-be/requirements.txt
Flask==3.0.0
pyjwt
accelerate
requests